[CAL-770] add video conferencing integration improvement
