+++ lib/integrations/calendar/interfaces/Calendar.ts
@@ -3,8 +3,8 @@ import { TFunction } from "next-i18next";
 
 import { PaymentInfo } from "@ee/lib/stripe/server";
 
+import { VideoCallData } from "@lib/integrations/videoConferencing/interfaces/VideoConferencing";
 import { Ensure } from "@lib/types/utils";
-import { VideoCallData } from "@lib/videoClient";
 
 import { NewCalendarEventType } from "../constants/types";
 import { ConferenceData } from "./GoogleCalendar";
