+++ lib/integrations/videoConferencing/constants/defaults.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_VIDEO_CONFERENCING_INTEGRATION_NAME = "";
